Remove donation stats service
